+++ fineract-provider/src/main/java/org/apache/fineract/accounting/journalentry/data/LoanDTO.java
@@ -45,5 +45,5 @@ public class LoanDTO {
     private boolean markedAsChargeOff;
     @Setter
     private boolean markedAsFraud;
-    private Integer chargeOffReasonCodeValue;
+    private Long chargeOffReasonCodeValue;
 }
